Use the RawNumber API to load numbers
This allows loading the (virtually) unlimited range of Python's integers
and floats (when using the Decimal data type). Since that the conversion
from the textual representation to the numeric is now entirely done by
the Python runtime, the flag precise_float does not makes sense anymore
and it's been removed.